analyzer: add plot_override_comment_rate_multiple_jira_tickets()
The most important data for deciding on urgency and
importance around particular CI instabilities is
with which rate a particular issue happens, resolved
over time. That is, not just the very recent past is
important, but the complete time evolution of an issue
since the beginning of time. This plot shows the
very same evolution for the top N JIRA tickets
referred to in override commands.